Init service candidates refactor
