+++ tests/unit/mocks/candidates.js
